chore(ci.yml): update Heroku deployment configuration to include environment variables and Procfile
This commit updates the Heroku deployment configuration to include the environment variables HD_SOCKET_URL and HD_NODE_ENV. It also adds a Procfile to specify the command to start the app in production. The configuration now uses the heroku_app_name, heroku_email, and heroku_api_key variables to deploy to Heroku.